Working With JavaScript guide.
Originally, this guide was called "AJAX on Rails," but really, it's
not just about AJAX. This was never finished, but it got accidentally
generated and Google found out about it.

In the meantime, all the guides were converted to markdown, as well.

So here's a new guide. It covers all of the built-in helpers that
use ajax requests, it covers CoffeeScript, and it covers UJS.