Fixed validation issue in Upgrading Ruby on Rails guide [ci skip]
In chapter 2.4 there was an underscore at the start of the attr_readonly method name,
this underscore was interpreted as the opening of the <em> tag but it wasn't closed.

Worst, the underscore in self.class.serialized_attributes (end of paragraph)
was interpreted as the closing of the <em> tag and not as the name of a method.