Fixed validation issue and titles in Migrations guide [ci skip]
Changed the title of the 3.5 chapter from "Using the change Method" to
"When to Use the change Method" since the chapter tells the migration
definitions supported by the change method. This fix the duplicate ID in the
document.

I've also changed some of the titles to respect the Guides guidelines.