AS guide: revises the description of #monday, and #sunday [ci skip]
The start of the week is a parameter nowadays, it is a little
confusing that we say here that we assume something about it.
Better to use "next" and "previous", which are clear regardless
of the concept of beginning of week.

Adds examples for the edge cases where the receiver is a
Monday or Sunday itself.