Updated 4.0 release notes
The release notes is changed to include only the major changes and point
to the changelogs for the complete list. This lets us talk more about the
features included in the upcoming release and make the guide more useful
and interesting than just a list of all changelogs.

This is still a WIP guide pending a lot of work.

[ci skip]