Revert "Detail middleware initialization points in the initialization guide [ci skip]"
I feel this is too detailed for the guide. Most of that is not public interface,
and if tomorrow something is added or removed in those chains we do not want
to come here to maintain it. Also, linking to a gist is not a good idea in a guide.

We could have a more broad description of what goes in what order, the most key and
public things.

This reverts commit 862a05edf3af128f25a87999936262b733ed781e.