Getting rid of a few other vestiges of rails.png.
We don't actually need a rails.png in the AP fixtures, the tests
that use it don't actually try to load the file.

We also don't need to get rid of it with the dummy reset either.

Finally, it's not needed in the sample application that's included
with the Rails Guides.