Follow up to 38ad5438 [ci skip]
This commit adds new lines between titles, paragraphs and code blocks
for an easier reading.

Moreover, in order to be consistent with the other guides, let's remove
the sharps on the right side of the different titles.

The "Running a Single Test" part was also explained after the "Testing
Active Record" one which shows how to run a single test under the scope
of a specific adapter.

Finally, let's explain the `-n` option to run a specific test method.