Update getting_started.md
It may be confusing to a Rails newbie why "Article" is suddenly being referred to with the "A" being capitalized. Thus far in the guide, the "a" has always been in lowercase. I think it is helpful to point out the reason for this change (especially since not using the capital "A" will result in an error). [ci skip]